+++ BackEndGoodLook/Controllers/UserController.cs
@@ -195,7 +195,7 @@ public class UserController : ControllerBase
     {
         return new UserSignDto()
         {
-            Id = (int)users.Id,
+            Id = users.Id,
             Name = users.Name,
             Email = users.Email,
             Password = users.Password,
